2.7.47 | /Controller/ApiEndpoints/ /Controller/mocking /Controller/Essentials/| Refactor. Rename Essentials dir to ApiEndpoints. Delete mocking dir
