+++ backend/src/Controller/ApiEndpoints/PostController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller\Essentials;
+namespace App\Controller\ApiEndpoints;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
